move HttpBody to cli/ops/fetch.rs
